+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- *BREAKING!* Switched to `derive_builder` for the builder, and renamed `AwakeHandle` to `KeepAwake`.
+
 ## 0.4.5 - 2023-11-08
 ### Changed
 - Migrate to winresource from winres.